feat: migration appcontroller and apploader classes to typescript
