+++ migration-newip-to-ts/src/interfaces/types.ts
@@ -37,3 +37,8 @@ export type UrlLoadOptions = {
 export type LoaderSource = {
   sources: string;
 };
+
+export interface IGetResp {
+  endpoint: string;
+  options?: LoaderSource | object;
+}
